#2 Finished BUY/SELL

+++ src/main/java/com/ibm/security/appscan/altoromutual/model/Account.java
@@ -57,7 +57,7 @@ public class Account {
 		return accountId;
 	}
 
-	public void setAccountId(int accountId) {
+	public void setAccountId(long accountId) {
 		this.accountId = accountId;
 	}
 
@@ -65,7 +65,7 @@ public class Account {
 		return balance;
 	}
 
-	public void setBalance(long balance) {
+	public void setBalance(double balance) {
 		this.balance = balance;
 	}
